requirements.txt
prefect==2.14.0
prefect-gcp==0.5.0
pandas==2.0.3
requests==2.31.0
google-cloud-bigquery==3.11.4
pandas-gbq==0.19.2
pyarrow==12.0.1
python-dotenv==1.0.0
pydantic==2.0.0
pytest==7.4.0
black==23.7.0